+++ main/java/Entity/Admin.java
@@ -1,0 +1,5 @@
+package Entity;
+
+public class Admin {
+
+}
