+++ main/java/Entity/Category.java
@@ -1,0 +1,8 @@
+package Entity;
+
+import Service.CategoryEditor;
+
+public class Category extends CategoryEditor {
+    String name;
+    String type; //season, casual
+}
