+++ main/java/Entity/Good.java
@@ -1,0 +1,9 @@
+package Entity;
+
+import Service.ItemEditor;
+
+public class Good extends ItemEditor {
+    String name;
+    int amount;
+    boolean availability;
+}
