+++ main/java/Entity/Info.java
@@ -1,0 +1,6 @@
+package Entity;
+
+import Service.PageEditor;
+
+public class Info extends PageEditor {
+}
