+++ main/java/Entity/SubCategory.java
@@ -1,0 +1,6 @@
+package Entity;
+
+public class SubCategory extends Category {
+    String name;
+    boolean isNeeded;
+}
