+++ main/java/Entity/SubGood.java
@@ -1,0 +1,6 @@
+package Entity;
+
+public class SubGood extends Good {
+    String name;
+    String type;
+}
