+++ main/java/Entity/User.java
@@ -1,0 +1,10 @@
+package Entity;
+
+import java.util.ArrayList;
+
+public class User {
+    String name;
+    int age;
+    boolean isOnline;
+    int moneySpent;
+}
