--- main/java/EssenceClasses/Category.java
@@ -1,6 +1,0 @@
-package EssenceClasses;
-
-import ServiceClasses.CategoryEditor;
-
-public class Category extends CategoryEditor {
-}
