--- main/java/EssenceClasses/Info.java
@@ -1,6 +1,0 @@
-package EssenceClasses;
-
-import ServiceClasses.PageEditor;
-
-public class Info extends PageEditor {
-}
