--- main/java/EssenceClasses/Item.java
@@ -1,6 +1,0 @@
-package EssenceClasses;
-
-import ServiceClasses.ItemEditor;
-
-public class Item extends ItemEditor {
-}
