--- main/java/EssenceClasses/Page.java
@@ -1,6 +1,0 @@
-package EssenceClasses;
-
-import ServiceClasses.PageEditor;
-
-public class Page extends PageEditor {
-}
