--- main/java/EssenceClasses/SubCategory.java
@@ -1,4 +1,0 @@
-package EssenceClasses;
-
-public class SubCategory extends Category {
-}
