--- main/java/EssenceClasses/SubItem.java
@@ -1,4 +1,0 @@
-package EssenceClasses;
-
-public class SubItem extends Item {
-}
