--- main/java/Interfaces/Action.java
@@ -1,4 +1,0 @@
-package Interfaces;
-
-public interface Action {
-}
