--- main/java/Interfaces/NotAvailable.java
@@ -1,4 +1,0 @@
-package Interfaces;
-
-public interface NotAvailable {
-}
