--- main/java/Interfaces/Promo.java
@@ -1,4 +1,0 @@
-package Interfaces;
-
-public interface Promo {
-}
