--- main/java/Interfaces/Sell.java
@@ -1,4 +1,0 @@
-package Interfaces;
-
-public interface Sell {
-}
