+++ main/java/Service/CategoryEditor.java
@@ -1,0 +1,4 @@
+package Service;
+
+public class CategoryEditor {
+}
