+++ main/java/Service/PageEditor.java
@@ -1,0 +1,4 @@
+package Service;
+
+public class PageEditor {
+}
