--- main/java/ServiceClasses/CategoryEditor.java
@@ -1,4 +1,0 @@
-package ServiceClasses;
-
-public class CategoryEditor {
-}
