--- main/java/ServiceClasses/ItemEditor.java
@@ -1,4 +1,0 @@
-package ServiceClasses;
-
-public class ItemEditor {
-}
