--- main/java/ServiceClasses/PageEditor.java
@@ -1,4 +1,0 @@
-package ServiceClasses;
-
-public class PageEditor {
-}
